Moved /home page content to main page and moved sections directory components to /components directory
